Refactoring keyboard handling
The dispatcher module will listen for key down event and ask the
keyboard module to dispatch them.
The keyboard module is now responsible for dispatching key events (only
the ones we want to detect).
This is to simplify and clean the key handling in dispatcher. The key
event being dispatched sets as context the current dom element (exactly
as if it was a keydown event) and passes the original event object so
that preventDefault and stopPropagation could also be executed.