TAB won't trigger the switch between blocks anymore
We decided to leave the TAB outside the logic of switch as every
browser has its own TAB implementation and we would probably interfere
with that (accessibility, focus on links, etc…). The switch is
responsability of arrow keys, despite pressing TAB could result to
focusing the next editable element. In the future we could have nested
list so TAB will be useful again :)