Add parser#isTextEndOfHost
Add methods to detect if the cursor is at the visible end of a block. These methods ignore whitespace in their calculations.
Also I fixed a bug in isEndOfHost which treated the offset of elements different than the offset of text nodes.
The specs now contain some tests that work directly with rangy and I added some helper methods to make it easier to test with real ranges, so this hopefully does not happen again.