Add #isCursor and #isSelection to Cursor and Selection
Make it as easy with instances of Cursor and Selection to detect their type as it is now with RangeContainer instances. Plus the use of instanceof is neither reliable nor useful outside of the EditableJS closure.

I also refactored the selectionWatcher#selectionChanged dispatcher to make it easier to understand and remove duplicate code.