Add rangy 1.3alpha to vendor
We still use rangy 1.2.3 but I wanted to add 1.3alpha so we can have a quick look at the newest available code. Some problems like expanding selections by characters or words are implemented in 1.3 so it is an option for us to try to switch or to check out how rangy solved problems that we still need to solve.