Simplify content#insertCharacter()
Inserting a <span> is not necessary. One can directly insert a TextNode as well. *duh*
This solves the bug we had previously where the selection was not correct after the manipulation if text on the start or end of a TextNode was selected. But at the price of the inserted TextNodes not being glued to their surrounding text. I added expectations to the spec so everyone can easily see what exactly happens.