Sort out content cleanup in default behaviors
Previously we were cleaning out content on focus and
blur despite what was happening, this resulted in bad
markup when merging. In my opinion a better approach
is to do the cleaning you need, only when you need it.
Thus, I did the cleaning in merge and split. Still we have
something when blur that is necessary as it has to
happen after leaving a block where we inserted a new
line; we will use the blur also for trimming traling spaces.