Experimental Selection#toggleLink()
This is an implementation that expands to the whole <a> tag if it is only partially selected. If multiple links are selected it will select the first tag. This is not really usable as of now since in a real use case the user will need the possibility to edit the link.