Add Cursor#save() and #restore()
Add RangeSaveRestore functionality directly to the Cursor class. To get this to work the cursor must also update its host if someone moves the cursor. For that I added the #setHost method that should be called whenever the host could have changed. I also store the host when saving a cursor, so it becomes possible to move the cursor to another block and then restore it to the previous block.

On top of that I added the convenience methods #moveAtBeginning() and #moveAtEnd() which I wanted to use in the specs. While at it I also refactored the #moveAfter() and #moveBefore() so they use the rangy methods and set the range more where one would expect it.