Add commented out spec
This spec fails as long as a saved range adds an invisible
non-breaking empty space. As of now this space is not ignored by
the extract text and highlight match methods. And removing the space
breaks the current merge implementation since the stored cursor position
will be remove during a merge without it.