Refactor error prone code
For some reason unbeknownst to me when converting the code
to browserify reassigning the variables which were also passed
as arguments started to change the arguments collection and the splicing
of the arguments would be different than before.

This behaviour is expected as arguments is not a normal array. The question
is why it didn't occur before browserifying the project...

To prevent this we convert the arguments into a proper array right
at the beginning now.