fix(clipboard): Blacklist pasted HTML elements
Background: rich text get pasted with their CSS <style> tag content,
since this <style> tag is not one of the tag we'd like to keep, we were
stripping the tag which kept its content and put it in <p>'s. This fixes that
by removing not only the <style> tag but also its content. This is made
configurable via `pastedHtmlRules.blacklistedHtmlElements`.